comment rating system

+++ app/model/services/RatingService.php
@@ -3,6 +3,8 @@
 namespace App\Model\Services;
 
 
+use App\Model\Entities\Comment;
+use App\Model\Entities\CommentRating;
 use App\Model\Entities\Quote;
 use App\Model\Entities\QuoteRating;
 use App\Model\Entities\User;
@@ -22,18 +24,66 @@ class RatingService extends Object
         $this->em = $entityManager;
     }
 
-    public function rate(Quote $quote, User $user, $positive)
+    public function rateComment(Comment $comment, User $user, $positive)
+    {
+        /** @var CommentRating $rating */
+        $rating = $this->em->getRepository(CommentRating::class)->findOneBy(['user' => $user, 'comment' => $comment]);
+        if ($rating) {
+            if ($rating->isPositive() && $positive || $rating->isNegative() && !$positive) { // user is removing his rating
+                $this->em->remove($rating);
+            } else { // user has changed his mind
+                $positive ? $rating->setPositive() : $rating->setNegative();
+            }
+        } else {
+            $rating = new CommentRating($user, $comment);
+            $positive ? $rating->setPositive() : $rating->setNegative();
+
+            $this->em->persist($rating);
+        }
+
+    }
+
+    public function rateQuote(Quote $quote, User $user, $positive)
     {
         /** @var QuoteRating $rating */
         $rating = $this->em->getRepository(QuoteRating::class)->findOneBy(['user' => $user, 'quote' => $quote]);
         if ($rating) {
-            $this->em->remove($rating);
+            if ($rating->isPositive() && $positive || $rating->isNegative() && !$positive) { // user is removing his rating
+                $this->em->remove($rating);
+            } else { // user has changed his mind
+                $positive ? $rating->setPositive() : $rating->setNegative();
+            }
         } else {
             $rating = new QuoteRating($user, $quote);
             $positive ? $rating->setPositive() : $rating->setNegative();
 
             $this->em->persist($rating);
         }
+    }
+
+    /**
+     * Comment rating algo.
+     *
+     * @param Comment $comment
+     */
+    public function updateCommentRating(Comment $comment)
+    {
+        $ups = $this->calculateCommentRatings($comment, CommentRating::POSITIVE);
+        $downs = $this->calculateCommentRatings($comment, CommentRating::NEGATIVE);
+        $comment->setRatingUp($ups);
+        $comment->setRatingDown($downs);
+    }
+
+    private function calculateCommentRatings(Comment $comment, $value)
+    {
+        $sum = 0;
+        /** @var CommentRating $r */
+        foreach ($comment->getRatings() as $r) {
+            if ($r->getValue() == $value) {
+                $sum++;
+            }
+        }
+        return $sum;
     }
 
     /**
@@ -49,6 +99,7 @@ class RatingService extends Object
         }
 
         $sum = 0;
+        /** @var QuoteRating $r */
         foreach ($quote->getRatings() as $r) {
             $sum += $r->getValue();
         }
